requirements.txt
Django==4.0.2
djangorestframework==3.13.1
flake8==4.0.1
pre-commit==2.17.0
requests==2.27.1
virtualenv==20.13.1